feat: refactor eraser utility functions and update CSS background color
