+++ app/globals.css
@@ -32,7 +32,7 @@
     }
 
     .dark {
-        --background: 240 10% 3.9%;
+        --background: hsl(0 0% 9%);
         --foreground: 0 0% 98%;
         --card: 240 10% 3.9%;
         --card-foreground: 0 0% 98%;
