+++ package.json
@@ -27,7 +27,7 @@
         "eslint-plugin-prettier": "^5.2.3",
         "konva": "^9.3.18",
         "lucide-react": "^0.474.0",
-        "next": "15.1.6",
+        "next": "^15.2.0",
         "next-themes": "^0.4.4",
         "prettier-plugin-tailwindcss": "^0.6.11",
         "react": "^19.0.0",
@@ -40,7 +40,7 @@
     },
     "devDependencies": {
         "@eslint/eslintrc": "^3.3.0",
-        "@next/eslint-plugin-next": "^15.1.7",
+        "@next/eslint-plugin-next": "^15.2.0",
         "@testing-library/dom": "^10.4.0",
         "@testing-library/jest-dom": "^6.6.3",
         "@testing-library/react": "^16.2.0",
